move context layers to own component
