+++ scripts/pages/flows.page.js
@@ -6,6 +6,7 @@ import FlowContentContainer from 'containers/flow-content.container';
 import SankeyContainer from 'containers/sankey.container';
 import ColumnsSelectorContainer from 'containers/columns-selector.container';
 import MapLayersContainer from 'containers/map-layers.container';
+import MapContextContainer from 'containers/map-context.container';
 import MapLegendContainer from 'containers/map-legend.container';
 import MapContainer from 'containers/map.container';
 import NavContainer from 'containers/nav.container';
@@ -60,6 +61,7 @@ new SankeyContainer(store);
 new ColumnsSelectorContainer(store);
 new MapContainer(store);
 new MapLayersContainer(store);
+new MapContextContainer(store);
 new MapLegendContainer(store);
 new NavContainer(store);
 new TitlebarContainer(store);
